+++ pkg/appmanager/appManager.go
@@ -127,9 +127,8 @@ type Manager struct {
 	manageIngressClassOnly bool
 	ingressClass           string
 	// Ingress SSL security Context
-	rsrcSSLCtxt     map[string]*v1.Secret
-	WatchedNS       WatchedNamespaces
-	RoutesProcessed RoutesMap
+	rsrcSSLCtxt map[string]*v1.Secret
+	WatchedNS   WatchedNamespaces
 	// AS3 Specific features that can be applied to a Route/Ingress
 	trustedCertsCfgmap string
 	intF5Res           InternalF5ResourcesGroup
@@ -213,8 +212,6 @@ const (
 	hubModeInterval = 30 * time.Second //Hubmode ConfigMap resync interval
 )
 
-var RoutesProcessed []*routeapi.Route
-
 // Create and return a new app manager that meets the Manager interface
 func NewManager(params *Params) *Manager {
 	vsQueue := workqueue.NewNamedRateLimitingQueue(
@@ -256,7 +253,6 @@ func NewManager(params *Params) *Manager {
 		rsrcSSLCtxt:            make(map[string]*v1.Secret),
 		trustedCertsCfgmap:     params.TrustedCertsCfgmap,
 		intF5Res:               make(map[string]InternalF5Resources),
-		RoutesProcessed:        make(RoutesMap),
 		dgPath:                 params.DgPath,
 		agRspChan:              params.AgRspChan,
 		processAgentLabels:     params.ProcessAgentLabels,
@@ -1918,27 +1914,32 @@ func (appMgr *Manager) syncRoutes(
 	// buffer to hold F5Resources till all routes are processed
 	bufferF5Res := InternalF5Resources{}
 
-	var routesProcessed []string
 	routePathMap := make(map[string]string)
 	for _, route := range routeByIndex {
 		if route.ObjectMeta.Namespace != sKey.Namespace {
 			continue
 		}
 
-		// Mark each resource  as it is already processed during init Time
-		// So that later the create event of the same resource will not processed, unnecessarily
-		appMgr.processedResources[prepareResourceKey(Routes, sKey.Namespace, route.Name)] = true
-
-		key := route.Spec.Host + route.Spec.Path
-		if host, ok := routePathMap[key]; ok {
-			if host == route.Spec.Host {
-				log.Debugf("[CORE] Route exist with same host: %v and path: %v", route.Spec.Host, route.Spec.Path)
-				continue
+		var key string
+		if route.Spec.Path == "/" || len(route.Spec.Path) == 0 {
+			key = route.Spec.Host
+		} else {
+			key = route.Spec.Host + route.Spec.Path
+		}
+		if _, ok := routePathMap[key]; ok {
+			if _, ok := appMgr.processedResources[prepareResourceKey(Routes, sKey.Namespace, route.Name)]; !ok {
+				log.Warningf("[CORE]  Route exist with same host: %v and path: %v", route.Spec.Host, route.Spec.Path)
+				// Adding the entry for resource so logs does not print repeatedly
+				// Putting the value as false so that route status does not get updated
+				appMgr.processedResources[prepareResourceKey(Routes, sKey.Namespace, route.Name)] = false
 			}
+			continue
 		} else {
 			routePathMap[key] = route.Spec.Host
 		}
-		routesProcessed = append(routesProcessed, route.ObjectMeta.Name)
+		// Mark each resource  as it is already processed during init Time
+		// So that later the create event of the same resource will not processed, unnecessarily
+		appMgr.processedResources[prepareResourceKey(Routes, sKey.Namespace, route.Name)] = true
 
 		//FIXME(kenr): why do we process services that aren't associated
 		//             with a route?
@@ -2107,12 +2108,6 @@ func (appMgr *Manager) syncRoutes(
 		updateDataGroupForABRoute(route, svcName, DEFAULT_PARTITION, sKey.Namespace, dgMap)
 
 		appMgr.processAS3SpecificFeatures(route, bufferF5Res)
-	}
-
-	if len(routesProcessed) != 0 {
-		appMgr.RoutesProcessed[sKey.Namespace] = routesProcessed
-	} else {
-		delete(appMgr.RoutesProcessed, sKey.Namespace)
 	}
 
 	// if buffer is updated then update the appMgr and stats
@@ -3317,6 +3312,24 @@ func (appMgr *Manager) exposeKubernetesService(
 
 func prepareResourceKey(kind, namespace, name string) string {
 	return kind + "_" + namespace + "/" + name
+}
+
+func getProcessedResources(processedResources map[string]bool, kind string) map[string][]string {
+	resourceMap := make(map[string][]string)
+	for k, v := range processedResources {
+		firstSplit := strings.Split(k, "_")
+		// considered only those resources which has value as true
+		if firstSplit[0] == kind && v {
+			secondSplit := strings.Split(firstSplit[1], "/")
+			if _, ok := resourceMap[secondSplit[0]]; ok {
+				resourceMap[secondSplit[0]] = append(resourceMap[secondSplit[0]], secondSplit[1])
+			} else {
+				resourceMap[secondSplit[0]] = []string{secondSplit[1]}
+			}
+
+		}
+	}
+	return resourceMap
 }
 
 func getIngressBackend(ing *v1beta1.Ingress) []string {