+++ docs/RELEASE-NOTES.rst
@@ -7,6 +7,7 @@ Added Functionality
 ```````````````````
 Enhancements
 ````````````
+* Added fix for processing oldest route when same host and path in routes
 
 2.7.1
 -------------